Modularized the environment and config file and all sensitive info now exists in .env which is added to gitignore
