+++ api/.env
@@ -1,2 +1,7 @@
 secret = 'what is a secret blud'
 BLOB_READ_WRITE_TOKEN="************"
+
+DB_USER=akifali
+DB_PASSWORD=5yCouf1k7DE5i1TI
+DB_CLUSTER=cluster0.7lc11ox
+DB_NAME=homease
